@@ -1,12 +1,60 @@
-import React from 'react';
+import { Card, CardHeader, CardTitle, CardDescription, CardContent, CardFooter } from "@/components/ui/card"
+import { Button } from "@/components/ui/button"
+import { Label } from "@/components/ui/label"
+import { Input } from "@/components/ui/input"
+import { Progress } from "@/components/ui/progress"
+import { Switch } from "@/components/ui/switch"
+import { ChromeIcon, AppleIcon } from "lucide-react"
 
-const Register: React.FC = () => {
+export default function Register() {
   return (
-    <div>
-      <h1>Register</h1>
-      <p>This is the Register component.</p>
+    <div className="flex items-center justify-center min-h-screen bg-gray-100">
+      <Card className="w-full max-w-md">
+        <CardHeader className="space-y-1">
+          <CardTitle className="text-2xl">Register</CardTitle>
+          <CardDescription>Enter your email and password to create your account</CardDescription>
+        </CardHeader>
+        <CardContent className="grid gap-4">
+          <div className="grid grid-cols-2 gap-6">
+            <Button variant="outline">
+              <ChromeIcon className="mr-2 h-4 w-4" />
+              Sign up with Google
+            </Button>
+            <Button variant="outline">
+              <AppleIcon className="mr-2 h-4 w-4" />
+              Sign up with Apple
+            </Button>
+          </div>
+          <div className="relative">
+            <div className="absolute inset-0 flex items-center">
+              <span className="w-full border-t" />
+            </div>
+            <div className="relative flex justify-center text-xs uppercase">
+              <span className="bg-background px-2 text-muted-foreground">Or continue with</span>
+            </div>
+          </div>
+          <div className="grid gap-2">
+            <Label htmlFor="memberName">Member Name</Label>
+            <Input id="memberName" placeholder="Enter your name" />
+          </div>
+          <div className="grid gap-2">
+            <Label htmlFor="email">Email Address</Label>
+            <Input id="email" type="email" placeholder="m@example.com" />
+          </div>
+          <div className="grid gap-2">
+            <Label htmlFor="password">Password</Label>
+            <Input id="password" type="password" />
+            <Progress value={25} className="h-2" />
+          </div>
+          <div className="flex items-center space-x-2">
+            <Switch id="terms" />
+            <Label htmlFor="terms">Accept terms of use</Label>
+          </div>
+        </CardContent>
+        <CardFooter>
+          <Button className="w-full">Register</Button>
+        </CardFooter>
+      </Card>
     </div>
   );
-};
-
-export default Register;
+}
